+++ apps/docs/CHANGELOG.md
@@ -5,5 +5,5 @@
 ### Patch Changes
 
 - Updated dependencies [125b9ff]
-  - @repo/piano@0.1.0
-  - @repo/react@0.1.0
+  - @music-ui/piano@0.1.0
+  - @music-ui/react@0.1.0